Zoom

+++ src/main/java/com/example/demo4/Observer.java
@@ -1,6 +1,6 @@
 package com.example.demo4;
 
 public interface Observer {
-    void update(int x,int y,int zoom);
+    void update(int x,int y,double zoom);
     void update(String imgUrl);
 }
